requirements.txt
PyYAML~=6.0.1
colorama~=0.4.6
pydantic~=1.10.2